Rewrite Root App component in TS and provide isDark context from it
